# api/docs/swagger.yaml
definitions:
...
  models.Top_Properties:
    properties:
      property_id:
        type: string
      tariff_name:
        type: string
    type: object
  models.Top_UpdateProperties:
    properties:
      end_date:
        type: string
      id:
        type: string
      property_id:
        type: string
      start_date:
        type: string
      tariff_name:
        type: string
    type: object
...
  tariff.CreateTariffReq:
    properties:
      days:
        type: integer
      name:
        type: string
      offers:
        type: string
      price:
        type: number
    type: object
  tariff.CreateTariffRes:
    properties:
      id:
        type: string
    type: object
  tariff.DeleteTariffRes:
    properties:
      success:
        type: boolean
    type: object
  tariff.GetAllTariffRes:
    properties:
      tariffs:
        items:
          $ref: '#/definitions/tariff.Tariff'
        type: array
    type: object
  tariff.GetTariffRes:
    properties:
      tariff:
        $ref: '#/definitions/tariff.Tariff'
    type: object
  tariff.Tariff:
    properties:
      created_at:
        type: string
      days:
        type: integer
      id:
        type: string
      name:
        type: string
      offers:
        type: string
      price:
        type: number
      updated_at:
        type: string
    type: object
  tariff.UpdateTariffReq:
    properties:
      days:
        type: integer
      id:
        type: string
      name:
        type: string
      offers:
        type: string
      price:
        type: number
    type: object
  tariff.UpdateTariffRes:
    properties:
      success:
        type: boolean
    type: object
  top_properties.CreateTopPropertyRes:
    properties:
      id:
        description: Top property ID
        type: string
    type: object
  top_properties.DeleteTopPropertyRes:
    properties:
      success:
        type: boolean
    type: object
  top_properties.GetAllTopPropertyRes:
    properties:
      top_properties:
        items:
          $ref: '#/definitions/top_properties.TopProperty'
        type: array
    type: object
  top_properties.GetTopPropertyRes:
    properties:
      top_property:
        $ref: '#/definitions/top_properties.TopProperty'
    type: object
  top_properties.TopProperty:
    properties:
      created_at:
        type: string
      end_date:
        type: string
      id:
        type: string
      property_id:
        type: string
      start_date:
        type: string
      tariff_name:
        type: string
      user_id:
        type: string
    type: object
  top_properties.UpdateTopPropertyRes:
    properties:
      success:
        type: boolean
    type: object
...
info:
  contact: {}
  description: This is a sample server for Api-gateway Service
  title: Api Gateway
  version: "1.0"
paths:
...
  /tarif/createtarif:
    post:
      consumes:
      - application/json
      description: Create a new tariff with specified details
      parameters:
      - description: Create Tariff Request
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/tariff.CreateTariffReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/tariff.CreateTariffRes'
        "400":
          description: Bad Request
          schema:
            type: string
      security:
      - ApiKeyAuth: []
      summary: Create a new tariff
      tags:
      - tariffs
  /tarif/deletetarif/{tarif_id}:
    delete:
      consumes:
      - application/json
      description: Delete a tariff by tariff ID
      parameters:
      - description: Tariff ID
        in: path
        name: tarif_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/tariff.DeleteTariffRes'
        "400":
          description: Bad Request
          schema:
            type: string
      security:
      - ApiKeyAuth: []
      summary: Delete tariff
      tags:
      - tariffs
  /tarif/getalltarif/{limit}/{page}:
    get:
      consumes:
      - application/json
      description: Retrieve a list of all tariffs with pagination
      parameters:
      - description: Limit
        in: path
        name: limit
        required: true
        type: integer
      - description: Page
        in: path
        name: page
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/tariff.GetAllTariffRes'
        "400":
          description: Bad Request
          schema:
            type: string
      security:
      - ApiKeyAuth: []
      summary: Get all tariffs
      tags:
      - tariffs
  /tarif/getbyidtarif/{tarif_id}:
    get:
      consumes:
      - application/json
      description: Retrieve tariff details by tariff ID
      parameters:
      - description: Tariff ID
        in: path
        name: tarif_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/tariff.GetTariffRes'
        "400":
          description: Bad Request
          schema:
            type: string
      security:
      - ApiKeyAuth: []
      summary: Get tariff by ID
      tags:
      - tariffs
  /tarif/updatetarif:
    put:
      consumes:
      - application/json
      description: Update an existing tariff with specified details
      parameters:
      - description: Update Tariff Request
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/tariff.UpdateTariffReq'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/tariff.UpdateTariffRes'
        "400":
          description: Bad Request
          schema:
            type: string
      security:
      - ApiKeyAuth: []
      summary: Update a tariff
      tags:
      - tariffs
  /topproperties/createtopproperties:
    post:
      consumes:
      - application/json
      description: Create a new top property with specified details
      parameters:
      - description: Top Properties
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/models.Top_Properties'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/top_properties.CreateTopPropertyRes'
        "400":
          description: Bad Request
          schema:
            type: string
      security:
      - ApiKeyAuth: []
      summary: Create top properties
      tags:
      - top_properties
  /topproperties/deletetopproperties/{top_properties_id}:
    delete:
      consumes:
      - application/json
      description: Delete a top property by its ID
      parameters:
      - description: Top Properties ID
        in: path
        name: top_properties_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/top_properties.DeleteTopPropertyRes'
        "400":
          description: Bad Request
          schema:
            type: string
      security:
      - ApiKeyAuth: []
      summary: Delete top properties
      tags:
      - top_properties
  /topproperties/getalltopproperties/{limit}/{page}:
    get:
      consumes:
      - application/json
      description: Retrieve a list of all top properties with pagination
      parameters:
      - description: Limit
        in: path
        name: limit
        required: true
        type: integer
      - description: Page
        in: path
        name: page
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/top_properties.GetAllTopPropertyRes'
        "400":
          description: Bad Request
          schema:
            type: string
      security:
      - ApiKeyAuth: []
      summary: Get all top properties
      tags:
      - top_properties
  /topproperties/getbyidtopproperties/{top_properties_id}:
    get:
      consumes:
      - application/json
      description: Retrieve top properties details by its ID
      parameters:
      - description: Top Properties ID
        in: path
        name: top_properties_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/top_properties.GetTopPropertyRes'
        "400":
          description: Bad Request
          schema:
            type: string
      security:
      - ApiKeyAuth: []
      summary: Get top properties by ID
      tags:
      - top_properties
  /topproperties/updatetopproperties:
    put:
      consumes:
      - application/json
      description: Update an existing top property by its ID
      parameters:
      - description: Update Top Properties
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/models.Top_UpdateProperties'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/top_properties.UpdateTopPropertyRes'
        "400":
          description: Bad Request
          schema:
            type: string
      security:
      - ApiKeyAuth: []
      summary: Update top properties
      tags:
      - top_properties
  /upload/imagesandvideo:
    post:
      consumes:
      - multipart/form-data
      description: Fayl yuklash uchun endpoint
      parameters:
      - description: Yuklanadigan fayl
        in: formData
        name: file
        required: true
        type: file
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: string
        "400":
          description: Bad Request
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      security:
      - ApiKeyAuth: []
      summary: Fayl yuklash
      tags:
      - Media
...
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"